Extract ES7 observable interoperability to separate class
