+++ src/observable.js
@@ -2,6 +2,7 @@ const {extend} = require('./utils/objects');
 const {VALUE, ERROR, ANY, END} = require('./constants');
 const {Dispatcher, callSubscriber} = require('./dispatcher');
 const {findByPred} = require('./utils/collections');
+const ESObservable = require('./es-observable');
 
 
 
@@ -101,22 +102,6 @@ extend(Observable.prototype, {
     return this._on(ANY, fn);
   },
 
-  subscribe(observer) {
-    let stream = this.takeErrors(1);
-    let fn = function(event) {
-      if (event.type === "value" && observer.next) {
-        observer.next(event.value);
-      } else if (event.type == "error" && observer.error) {
-        observer.error(event.value);
-      } else if (event.type === "end" && observer.complete) {
-        observer.complete(event.value);
-      }
-    }
-
-    stream.onAny(fn);
-    return () => stream.offAny(fn);
-  },
-
   offValue(fn) {
     return this._off(VALUE, fn);
   },
@@ -182,7 +167,9 @@ extend(Observable.prototype, {
 
 });
 
-Observable.prototype[require('./utils/symbol')('observable')] = function() { return this; }
+Observable.prototype[require('./utils/symbol')('observable')] = function() {
+  return new ESObservable(this);
+}
 
 // extend() can't handle `toString` in IE8
 Observable.prototype.toString = function() {
